Add files via upload
Multiple screenshots of dashboard are uploaded. 
In image 1.PNG
Whole dashboard can be seen with no filter. 

In image 2.PNG
Human resource department is selected and attrition data of HR is only showing on different KPIs. 

In image 3.PNG
Male employees filter is applied so that dashboard will only shows the attrition data of males only on different KPIs. 

In image 4.PNG
Research and Development department and Male filter is turned on so it shows only attrition data of a R&D department of males only.